requirements.txt
streamlit
openai
cohere
python-dotenv
pyaudio
whisper
pyttsx3



